Added Ultrasonic support, it now walks!
Added ultrasonic sensor support (just reading data, no behavior based on that for now);
Changed the walk routine for a better gait: it now walks! :D
Added a macro for step size and changed the loop delay (to 0, for now) for smoother and faster movement!
It now sets the shoulder servos at 0 on setup, so that it can actually work;
Lowered general robot for better stability